xiugai

+++ usecase/src/main/java/data/usecase/DeleteSynParameter.java
@@ -1,5 +1,7 @@
 package data.usecase;
 
+import data.entity.SynParameter;
+import data.usecase.exception.NotExistException;
 import data.usecase.port.SynParmeterRepository;
 import data.vo.id.AppId;
 import lombok.AllArgsConstructor;
@@ -13,7 +15,12 @@ import lombok.Setter;
 @Setter
 public class DeleteSynParameter {
     private SynParmeterRepository synParmeterRepository;
-    public void execut(AppId appid){
-        synParmeterRepository.deletePerameter(appid);
+    public void execut(AppId appid) throws NotExistException{
+        if(synParmeterRepository.findParameter(SynParameter.builder().synParameterId(appid).build())!=null){
+            synParmeterRepository.deletePerameter(appid);
+        }else {
+            throw new NotExistException(1000,"任务不存在");
+        }
+
     }
 }
